backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
sqlalchemy==2.0.*
psycopg2-binary==2.9.*
python-jose[cryptography]==3.3.*
passlib[bcrypt]==1.7.*
bcrypt==4.0.1
pydantic[email]==2.10.*
requests==2.32.*
